+++ recipes/default.rb
@@ -9,6 +9,7 @@ node['winsw']['service'].each do |service_name, service_info|
     executable service_info['executable']
     args service_info['args']
     startargs service_info['start_args']
+    startmode service_info['startmode']
     env_variables service_info['env_variables']
     options service_info['options']
     extensions service_info['extensions']